wddlock.c: Don't use _al_wd_dirty_lines if it's null
Related to issue 359: Program exits unexpectently after minimizing
and maximizing too fast.

I should find the cause of _al_wd_dirty_lines = NULL, but I wasn't able
to reproduce this bug, so at the moment this is the best I can do to avoid
a crash. (I suppose that it's related to a temporary invalid state because
thread sync issues when the screen/window changes its size.)