Fix selection tool behavior when Shift key is used
There was a conflict between Shift key as a modifier to Add selection and
to convert the selection box with square-aspect. Now tools::Controllers
can process key down/up events, so the square aspect is used only if
TwoPointsController receives a Shift key event (in the middle of the
tool loop).