Fix tooltips visibility (fix #770)
TooltipManager uses filters to intercept a kMouseEnterMessage, if we
remove this kind of message for the TooltipManager itself (in sendMessage),
we never get that message so we never show the tooltip.

This problem was introduced in d1799c037abc011e79663892f52e0e2e0b5c5334
to fix a problem with comboboxes.