Add option to use a deprecated glibmethod
To have at least backwardcompatibility with ubuntu 15.10 a cmake option was added to make use of a deprecated glibmm method. Further testing ist need to see how far this backward compatibility works, as of know ubuntu 12.04 does not seem to work. Further an unused variable was removed, as well as a tab from src/she/CMakeLists.txt so it has no mixed intending.