Fix macOS issue: NSSavePanel is still open/visible after it's used
If we use the [NSSavePanel savePanel] singleton, after the file is
selected, we can Cmd+Tab to switch to another app and go back to our app
and the NSSavePanel will be visible again. With this patch we ensure 1)
we create our own panel, and 2) after selecting the filename, the panel
is destroyed/freed (so it will not be shown again when we switch between
apps).