Fix crashes using certain function w/Undo History open
Some commands were calling transaction.commit() in the non-main/UI
thread (e.g. SpriteSizeCommand). That commit() called
DocumentUndo::add() which generated a
DocumentUndoObserver::onAddUndoState() and it updated the Undo History
window UI. This generated a lot of racing conditions in the UI module
and possibilities of crashes if the Undo History window was visible.