Use the pointer API for mouse events too
Mouse messages and pointer messages do not go well together.
If we don't use the pointer API for mouse messages, we start
receiving random mouse messages when a pointer is being used
(this is increased when a modifier key is pressed).

With this patch we start using the pointer API for mouse
messages and remove our workaround to avoid these
random messages.

This might be a possible solution for issues like:
https://community.aseprite.org/t/all-the-tools-stop-working/738
https://community.aseprite.org/t/after-using-the-snipping-tool-on-win-10-the-mouse-pointer-disappears/906
http://steamcommunity.com/app/431730/discussions/0/1485487749769857401/