Undo resize operation correctly in File > Export w/Resize != 100% and without undo history
If we open a new image (there is no undo history) and use File > Export
with a Resize parameter, the resize command must be undone, but the
UndoState to return will be nullptr (nullptr is before the first UndoState,
which is the resize operation for the export operation). So without this
fix we just resize and cannot go back before the resize.

This regression was introduced in a4d1833a9c37aa95c67185ee8632bd4b9f8d84fc