Fix problems changing HSV/HSL color sliders with mouse wheel
When we change HSV/HSL sliders, the RGB color is calculated and all
sliders are re-updated from that color. As HSV/HSL are floating point
values, there might be a small adjustment in the 0-100 conversion for
the sliders, so we have to lock the changed slider to keep its
original value set from the mouse wheel.

Fix: https://community.aseprite.org/t/1411