Fix problems assigning different but similar modifiers to mouse wheel actions
For example, before this change we were depending on the order of the
wheel actions. If an action were associated to <Ctrl> and other one to
<Ctrl+Shift>, when we pressed <Ctrl+Shift> we weren't returning the
action with most modifiers pressed (in this case <Ctrl+Shift>), but
instead the one that was checked first in the for loop (which could
be the one with <Ctrl>).