Fix gradient origin when the kFromCenter is not used (regression introduced in 250d40c0f3615db2eecc04d038da1ab52eeae265)
When the kFromCenter is enabled, we have to use the middle-point of
the two points controller as the origin of the floodfill algorithm,
but when kFromCenter is not enabled, we use the first point.  This is
the only way to make both cases work well in such a way that the
origin of the floodfill is not displaced when the mouse is moved.