Fix losing sprite center when we change zoom
This issue is pretty common when we change the zoom and the cursor is
outside the canvas, we lost the center of the sprite or the canvas
goes more far and far away. I've seen this problem on streamers and
myself using the program. Hoping to do a little of improvement in this
area with this patch.