Restore the default brush center to w/2 and h/2
This generated a lot of controversy. The new center ((w-1)/2
, (h-1)/2) was useful for scripts, but not too useful for right-handed
users that prefer the most bigger part of even-sized brushes to be in
the top-left edge (instead of bottom-right edge which was the new
behavior).

With this patch we revert the changes, and we'll see if we can add a
new option to change the default brush center which might be useful
for left-handed users.