Fix flicker during playback when the backup thread starts working
This includes a new version of RWLock that allows several
reader locks + a weak lock (the weak lock is used in the
backup thread). In this way the backup thread is not stopped
or canceled when we have just to read the sprite to render it
on the editor.

Regression introduced in 165940164218d8ad10c43f39d08f0559e859098d