Fix tab style when the sprite defined color is white
Now the text is painted with a negative black or white color depending
on the defined color in the sprite user data. Before this patch if the
tab was painted white (and the tab text in the theme was defined as
white) we weren't able to see the sprite name in the tab at all.