Color Sort For Hue, Alpha
Removes switch case with fall through. Removes nested switch cases. Refactors color sorting method to 1. push zero alpha colors to the front of the palette; 2. resort to value as a criterion for gray colors; 3. use saturation and value as back up comparisons for each other for equivalent values; 4. approximate gamma-to-linear when perceptual lightness is chosen.

Partial fix for #2901.