Fix crash on set_native_cursor_on_all_displays
The intention of this fix is to solve a crash reported on ticket 2674 (file name: Aseprite-crsh-1.3.beta6-x64.dmp)
According to the .dmp file an 'access violation reading location 0xFFFFFFFFFFFFFFFF' happens on line 118 in 'system.cpp', function 'set_native_cursor_on_all_displays(...)'
The source code line where the program crashed was:

118  display = display->parentDisplay()

This line is inside a 'while' loop which checks if 'display' is 'nullptr'. So, I guess, 'display' pointer was destroyed in other parallel thread.