Fix WheelAction::ToolSameGroup/ToolOtherGroup in DraggingValueState (fix #3459)
These actions weren't working correctly because we have to take
special care when we change the tool group, i.e. the "initial
position" used to calculate the delta from "mouse pos - initial pos"
must be updated to reset the delta to change tools inside the same
tool (this is the new m_initialPosSameGroup member variable).