Minor refactor of the OctreeMap class
Removed unnecessary code related to member variable m_maskColor of 'OctreeMap'.
The constructor 'OctreeMap(sprite)' is included to simplify the selection of 'm_maskIndex' according to the properties of the sprite (according to the format and transparent index), in addition to deciding if the mask color should be included in the final palette.