Destroy overlay in ~DraggableWidget destructor
There is an issue where we can drag a recent item and load a file
(using Alt+F or Ctrl+O) while dragging the recent item. With this
commit the overlay is destroyed after the file is open (without this
the overlay is kept floating).

Probably related to #3504 but we're not sure.