Fix brush and pasted image preview on empty cels (fix #3882)
Fix regression introduced in 24846eae10c7340a4e7f103b15494ea8338700a9
using the new RenderPlan structure. We have to include empty/nullptr
cels in the plan to paint the preview image in the selected
layer/frame (the preview image is the one used for the brush preview,
and for the transformation preview when we paste the clipboard content
on the canvas).