Fix crash when loading an extension with invalid dithering matrices and the console must be shown (fix #3914)
Before this fix, when installing dithering matrices if Aseprite couldn't find the file of some matrix described in the json, Aseprite would crash (this happened during the installation of an erroneous dithering matrices extension, and after every reboot of Aseprite).
The cause of the crash was the absence of the MainWindow instance during the ContextBar creation. When an error occurs, the console is called, but since MainWindows is not yet available, aseprite crashes.