Fix removing all frames results in a transparent background layer
Currently if we select all frames (as shown in
https://github.com/aseprite/aseprite/issues/3551), and we will remove
all of them - it also removes cel inside the primary frame, resulting in
a transparent background (again, as shown in issue#3551).

This patch keeps the frame number inside the cel that got deleted while
being on 0th frame, and restores it's frame number when we redo the
frame. This is needed since we don't restore the full cel in
cmd::ClearCel as we do in cmd::RemoveCel, we still use the same cel
before deletion.