Update INSTALL.md for windows compilation such that a misleading architecture option is avoided when calling the developer command prompt
As it exists today, the provided `call "...VsDevCmd.bat" -arch=x64` command will output the developer command prompt at startup then close itself after a few seconds. What's going on, is the command prompt is silently failing, because it doesn't recognize the "x64" value for the -arch argument. 

This change makes the example `call` have a generic value for target architecutre, and links to the documentation where the valid values can be reviewed.