[i18n] Clone and add "strings" repo in "data/strings.git" to get updated translations
As we've moved all the translations to the strings
repo (06a852d1d4d53e8ebfbb200bce2c2d9f9a2212bc), to facilitate the
i18n work now we load translations from "data/strings" and from
"data/strings.git" folders, where "strings.git" is a clone of the
strings repo (https://github.com/aseprite/strings.git) in
"build/bin/data/strings.git".

This clone is executed automatically in the cmake configuration stage
so it's transparent for the developer and the result is like having
the translations available in the same "aseprite" repo.