Fix enabling commands when they need to write a document in certain cases (fix #4310)
In certain situations a script creating a transaction could run
certain commands (or functions running commands) that will ask if the
command is enabled before executing it (Command::executeCommand()).

ContextFlags::update() is the function that checks if we can upgrade a
reader lock to a writer, which means that we can write the
document. The new fix in laf's RWLock::canWriteLockFromRead() makes
this condition work correctly.