Fix inactive window passes events to window below (#4561, #4835)
Prior to this fix, if all of the following conditions were met:
+ While a modal window (such as the Preferences dialog) was open.
+ 'Multiple window UI' was enabled.
+ A non-modal window (such as the Preview window) overlapped the modal
  window.
+ The non-modal window was selected, so this window is on top of
  the modal window.
Caused mouse clicks and movements within the overlay area of ​​the
non-modal window to be incorrectly passed to the window below, i.e.
the modal window.
Instead, it would be expected that no action would be taken.

@dacap note: At the moment this fix only works on macOS because the
os::WindowSpec::modal() parameter is only used on macOS.